Initial draft
Realised I got the puzzle logic wrong the first time, I thought the puzzle was asking for numbers made of ANY repeating sequence, not that first half = last half. So made that much more complicated that it needed to be, but hey at least I learned alot of c++ parsing so it was useful.